requirements.txt
Django==4.2.5
djangorestframework==3.15.2
